fix(security): close open CodeQL/Semgrep alerts, harden extract_ugc_features input, audit Scorecard (#1243) (#1299)
* fix(security): widen operands before multiplication in convolve/psnr/moment (CodeQL x6)

* fix(ai): validate dimensions and paths before subprocess run in extract_ugc_features

* fix(security): widen integer index operands in convolve/moment/psnr preserving float SIMD bit-exactness

* docs(security): document Scorecard and remaining code-scanning alerts

* docs(security): correct the CodeQL cast-widening record and alert numbering

The first draft described a (double) widening of the float products and
claimed alerts 922-928 closed; the landed change widens only the integer
index operands (numerically inert) and those alert numbers are Meson probe
files, not the multiplication rule. Record the real alert history
(30/31/33/706/707/708, maintainer-dismissed false positives kept as-is for
ADR-0138 / ADR-0179 bit-exactness), the master CodeQL re-run
(run 33920556824), and mark the osv-scanner ignore as a maintainer decision.

* fix(cli): suppress the cppcheck false positive on the variadic usage() overload

The required Cppcheck check has been failing this PR for three merge-train
rounds on a single finding:

  core/tools/cli_parse.cpp:325  wrongPrintfScanfArgNum (CWE-685)
  'fprintf format string requires 1 parameter but 2 are given.'

It is a false positive. This PR split the old single variadic template into a
zero-argument overload and a one-or-more-argument template. In the template,
`reason` is a runtime format string and the arguments arrive as a parameter
pack, so cppcheck can resolve neither the conversion count in `reason` nor
`sizeof...(Rest)` and reports a fixed mismatch regardless of the call site.

Suppressed inline with the justification next to the code, matching the
convention already used in core/src/libvmaf.c, feature_collector.c and
integer_adm.c. The zero-argument overload is unaffected: it uses fputs and
never treats `reason` as a format string, which is why the split was made in
the first place.

Verified locally:
  cppcheck --enable=warning,performance,portability --inline-suppr \
    --suppressions-list=.cppcheck-suppressions.txt --error-exitcode=1 \
    core/tools/cli_parse.cpp   -> exit 0
  ninja tools/vmaf             -> links clean

CI runs cppcheck without --enable=information, so the suppression cannot
itself fail as an unmatchedSuppression on a newer cppcheck (local 2.21.1 does
not raise the finding at all; the CI image's older version does).


---------

+++ core/src/feature/iqa/convolve.c
@@ -154,7 +154,7 @@ static void iqa_convolve_vertical_pass(const float *img_cache, int w, const stru
             for (int v = -vc; v <= vc - kh_even; ++v, ++k_offset) {
                 sum += img_cache[img_offset + (ptrdiff_t)v * w] * k->kernel_v[k_offset];
             }
-            dst[y * dst_w + x] = (float)(sum * scale);
+            dst[(ptrdiff_t)y * dst_w + x] = (float)(sum * scale);
         }
     }
 }
